requirements.txt
appdirs==1.4.3
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
cssselect==1.1.0
cycler==0.10.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.6
django-heroku==0.3.1
filelock==3.0.12
GetOldTweets3==0.0.11
gunicorn==20.0.4
idna==2.9
kiwisolver==1.2.0
lxml==4.5.0
matplotlib==3.2.1
numpy==1.18.3
oauthlib==3.1.0
Pillow==7.1.1
psycopg2==2.8.5
pyparsing==2.4.7
pyquery==1.4.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
tweepy==3.8.0
urllib3==1.25.8
vaderSentiment==3.2.1
virtualenv==20.0.18
virtualenv-clone==0.5.4
whitenoise==5.0.1
wordcloud==1.6.0
tweet-preprocessor==0.4.0
